[BE] refactor: group api 리팩토링
- group api 리팩토링
