+++ server/src/domains/group/api/GroupController.ts
@@ -17,7 +17,8 @@ import { Inject, Service } from 'typedi';
 import { GroupService } from '../service/GroupService';
 import { AlarmService } from '@domains/alarm/service/AlarmService';
 
-import { GroupDetailDto, GroupParam } from '../dto/groupDto';
+import { GroupDetailDto } from '../dto/GroupDetailDto';
+import { GroupParam } from '../dto/GroupParam';
 import { FilterdPageGroupDto } from '../dto/FilterdGroupDto';
 import { CreateGroupDto } from '../dto/CreateGroupDto';
 import { GroupActivityDto } from '../dto/GroupActivityDto';
@@ -154,6 +155,6 @@ export class GroupController {
   @OpenAPI({ summary: '그룹 정보를 가져오는 API' })
   @ResponseSchema(GroupDetailDto, { description: '그룹 정보 조회 완료' })
   async getGroupData(@Param('gid') gid: number) {
-    return await this.groupService.getGroupDetails(gid);
+    return await this.groupService.getGroupDetail(gid);
   }
 }
